weddingapp/apps/story/migrations/0002_auto_20150815_1712.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import models, migrations
import ckeditor.fields


class Migration(migrations.Migration):

    dependencies = [
        ('story', '0001_initial'),
    ]

    operations = [
        migrations.AlterField(
            model_name='story',
            name='story_body',
            field=ckeditor.fields.RichTextField(),
        ),
    ]
